feat(core): implement automatic room deletion based on inactivity
Introduce a new feature to automatically delete conversation rooms if they have not been
used for a specified period, defined by the `autoDeleteTimeout` configuration setting.
This change helps to keep the database tidy by removing unused rooms on a regular basis.

The room deletion process is executed in a background interval task set to run every
half hour. A new function `updateChatTime` has been added to update the last activity
time of a room whenever a chat occurs. Rooms are considered for deletion if their
`updatedTime` stamp is older than the `autoDeleteTimeout` value.
```