i18n(core): update config message
The 'thinking message' middleware has been removed from the core package, as it is no longer
necessary for the functionality provided by ChatLuna. This change includes the deletion of the
thinking message configuration and middleware implementation, as well as any related documentation.

BREAKING CHANGE: Middleware for sending 'thinking messages' has been removed. If your bot
relies on this functionality, you will need to implement it manually.