ci: update yarn install command in GitHub Actions workflow
The change modifies the 'Install dependencies' step in the GitHub Actions workflow
to use 'yarn' instead of 'yarn install'. This adjustment is likely to improve the
efficiency of the workflow by leveraging the global installation of Yarn where
available, rather than explicitly installing packages that are already present in
the global cache.