ci: update actions/cache version from v4 to v3
The workflow is modified to use actions/cache@v3 instead of v4 for cache management in the Yarn-related steps. This change affects the following jobs:
- Yarn Cache Directory (twice)
- Node 20 Build (twice)
- ESLint Build (twice)

This update might be necessary due to changes in the cache mechanism or to align with a newer, stable version of the GitHub Actions cache action.